roachtest: extract function to parse JUnit test results
Release note: None
